requirements.txt
# requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.6
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
streamlit>=1.28.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0